+++ src/main/java/no/digipost/api/client/util/DigipostApiMock.java
@@ -20,33 +20,52 @@ import io.undertow.server.HttpServerExchange;
 import io.undertow.server.handlers.BlockingHandler;
 import io.undertow.util.HeaderMap;
 import io.undertow.util.HttpString;
-import no.digipost.api.client.ApiServiceMock;
-import no.digipost.api.client.ApiServiceMock.Method;
-import no.digipost.api.client.ApiServiceMock.RequestsAndResponses;
 import no.digipost.api.client.Headers;
+import no.digipost.api.client.errorhandling.ErrorCode;
 import no.digipost.api.client.representations.*;
 import no.digipost.api.client.representations.sender.SenderFeature;
 import no.digipost.api.client.representations.sender.SenderInformation;
 import no.digipost.api.client.representations.sender.SenderStatus;
 import org.apache.commons.io.IOUtils;
 import org.apache.commons.io.output.ByteArrayOutputStream;
+import org.apache.commons.lang3.EnumUtils;
 import org.apache.http.client.methods.CloseableHttpResponse;
+import org.apache.http.entity.ByteArrayEntity;
 import org.bouncycastle.crypto.digests.SHA256Digest;
+import org.bouncycastle.openssl.jcajce.JcaPEMWriter;
 import org.bouncycastle.util.encoders.Base64;
+import org.joda.time.DateTime;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import javax.xml.bind.JAXB;
 import java.io.ByteArrayInputStream;
+import java.io.IOException;
+import java.io.OutputStreamWriter;
+import java.io.Writer;
+import java.net.ConnectException;
 import java.security.*;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 import java.util.concurrent.BlockingDeque;
+import java.util.concurrent.ConcurrentLinkedQueue;
 import java.util.concurrent.LinkedBlockingDeque;
+
+import static java.lang.Integer.parseInt;
+import static no.digipost.api.client.representations.MessageStatus.COMPLETE;
+import static org.apache.http.HttpStatus.SC_OK;
 import static org.joda.time.DateTime.*;
 
 public class DigipostApiMock implements HttpHandler {
+
+	public enum Method {
+		SEND_MULTIPART_MESSAGE,
+		GET_PRINT_KEY,
+		GET_SENDER_INFORMATION,
+		GET_CONTENT,
+		MULTIPART_MESSAGE,
+		GET_DOCUMENTS_EVENTS,
+		GET_DOCUMENT_STATUS
+	}
 
 	private static final Logger LOG = LoggerFactory.getLogger(DigipostApiMock.class);
 
@@ -84,8 +103,6 @@ public class DigipostApiMock implements HttpHandler {
 			"6vYEpI930hzj81irYEPyRvoDAaEv2Zc=\n" +
 			"-----END CERTIFICATE-----";
 
-
-	private byte[] stubbedResponse;
 	private int port;
 	private Undertow server;
 	private final BlockingDeque<ReceivedRequest> receivedRequests = new LinkedBlockingDeque<>();
@@ -94,15 +111,15 @@ public class DigipostApiMock implements HttpHandler {
 
 	@Override
 	public void handleRequest(HttpServerExchange httpContext) throws Exception {
-		String httpResponse = "200";
+		int httpResponse = 200;
 		HttpString method = httpContext.getRequestMethod();
+		String requestPath = httpContext.getRequestPath().toLowerCase();
 
 		ByteArrayOutputStream bao = new ByteArrayOutputStream();
 
 		if(method.equals(new HttpString("POST"))){
 			String multipart = new String(IOUtils.toByteArray(httpContext.getInputStream()));
-			String messageString = multipart.substring(multipart.indexOf("<?xml"));
-			messageString = messageString.substring(0, messageString.indexOf("</message>") + 10);
+			String messageString = getMessageFromXMLString(multipart);
 
 			ByteArrayInputStream messageStream = new ByteArrayInputStream(messageString.getBytes());
 
@@ -111,20 +128,19 @@ public class DigipostApiMock implements HttpHandler {
 			RequestsAndResponses requestsAndResponses = this.requestsAndResponsesMap.get(Method.MULTIPART_MESSAGE);
 			CloseableHttpResponse response = requestsAndResponses.getResponse(message.primaryDocument.subject);
 
-			httpContext.setStatusCode(response.getStatusLine().getStatusCode());
-			httpResponse = response.getStatusLine().getStatusCode() + "";
+			httpResponse = response.getStatusLine().getStatusCode();
 			response.getEntity().writeTo(bao);
 
-			requestsAndResponses.addRequest(new ApiServiceMock.DigipostRequest(message, getContentparts(multipart)));
+			requestsAndResponses.addRequest(new DigipostRequest(message, getContentparts(multipart)));
 
-		} else if(httpContext.getRequestPath().toLowerCase().equals("/printkey")) {
+		} else if(requestPath.equals("/printkey")) {
 			CloseableHttpResponse response = requestsAndResponsesMap.get(Method.GET_PRINT_KEY).getResponse(httpContext.getRequestPath());
 			EncryptionKey encryptionKey = new EncryptionKey();
 			encryptionKey.setKeyId(PRINT_ID);
 			encryptionKey.setValue(PRINT_KEY);
 			JAXB.marshal(encryptionKey, bao);
 
-		} else if(httpContext.getRequestPath().toLowerCase().equals("/getsenderinformation")) {
+		} else if(requestPath.equals("/getsenderinformation")) {
 			CloseableHttpResponse response = requestsAndResponsesMap.get(Method.GET_SENDER_INFORMATION).getResponse(httpContext.getRequestPath());
 			if(response == null) {
 				List<SenderFeature> senderFeatures = new ArrayList<>();
@@ -134,24 +150,21 @@ public class DigipostApiMock implements HttpHandler {
 				senderFeatures.add(SenderFeature.PRINTVALIDATION_PDFVERSION);
 				JAXB.marshal(new SenderInformation(9999L, SenderStatus.VALID_SENDER, senderFeatures), bao);
 			} else {
-				httpContext.setStatusCode(response.getStatusLine().getStatusCode());
-				httpResponse = response.getStatusLine().getStatusCode() + "";
+				httpResponse = response.getStatusLine().getStatusCode();
 				response.getEntity().writeTo(bao);
 			}
 
-		} else if(httpContext.getRequestPath().toLowerCase().equals("/getdocumentstatus")) {
+		} else if(requestPath.equals("/getdocumentstatus")) {
 			CloseableHttpResponse response = requestsAndResponsesMap.get(Method.GET_DOCUMENT_STATUS).getResponse(httpContext.getRequestPath());
-			httpContext.setStatusCode(response.getStatusLine().getStatusCode());
-			httpResponse = response.getStatusLine().getStatusCode() + "";
+			httpResponse = response.getStatusLine().getStatusCode();
 			response.getEntity().writeTo(bao);
 
-		} else if(httpContext.getRequestPath().toLowerCase().equals("/getdocumentevents")) {
+		} else if(requestPath.equals("/getdocumentevents")) {
 			CloseableHttpResponse response = requestsAndResponsesMap.get(Method.GET_DOCUMENTS_EVENTS).getResponse(httpContext.getRequestPath());
-			httpContext.setStatusCode(response.getStatusLine().getStatusCode());
-			httpResponse = response.getStatusLine().getStatusCode() + "";
+			httpResponse = response.getStatusLine().getStatusCode();
 			response.getEntity().writeTo(bao);
 
-		} else if(httpContext.getRequestPath().toLowerCase().equals("/")){
+		} else if(requestPath.equals("/")){
 			JAXB.marshal(new EntryPoint(certificate, new Link(Relation.CREATE_MESSAGE, new DigipostUri("http://localhost:9999/create")),
 					new Link(Relation.GET_PRINT_ENCRYPTION_KEY, new DigipostUri("http://localhost:9999/printkey")),
 					new Link(Relation.GET_SENDER_INFORMATION, new DigipostUri("http://localhost:9999/getsenderinformation")),
@@ -159,26 +172,31 @@ public class DigipostApiMock implements HttpHandler {
 
 		} else {
 			CloseableHttpResponse response = requestsAndResponsesMap.get(Method.GET_CONTENT).getResponse(httpContext.getRequestPath());
-			httpContext.setStatusCode(response.getStatusLine().getStatusCode());
-			httpResponse = response.getStatusLine().getStatusCode() + "";
+			httpResponse = response.getStatusLine().getStatusCode();
 			response.getEntity().writeTo(bao);
 		}
 
 		byte[] bytes = bao.toByteArray();
 		HeaderMap responseHeaders = httpContext.getResponseHeaders();
+		
 		String dateOnRFC1123Format = DateUtils.formatDate(now());
-
-		responseHeaders.add(new HttpString("Date"), dateOnRFC1123Format);
 		String xContentSHA256 = generateXContentSHA256(bytes);
-		responseHeaders.add(new HttpString(Headers.X_Content_SHA256), xContentSHA256);
-
 		String signature =  httpResponse + "\n" + httpContext.getRequestPath().toLowerCase() + "\n" +
 				"date: " + dateOnRFC1123Format + "\n" + "x-content-sha256: " + xContentSHA256 + "\n";
 
+		responseHeaders.add(new HttpString("Date"), dateOnRFC1123Format);
+		responseHeaders.add(new HttpString(Headers.X_Content_SHA256), xContentSHA256);
 		responseHeaders.add(new HttpString(Headers.X_Digipost_Signature), generateXDigipostSignature(keyPair, signature));
+
+		httpContext.setStatusCode(httpResponse);
 
 		httpContext.startBlocking();
 		IOUtils.copy(new ByteArrayInputStream(bytes), httpContext.getOutputStream());
+	}
+
+	private static String getMessageFromXMLString(String multipart) throws IOException {
+		String messageString = multipart.substring(multipart.indexOf("<?xml"));
+		return messageString.substring(0, messageString.indexOf("</message>") + 10);
 	}
 
 	private static String generateXContentSHA256(byte[] bytes){
@@ -205,6 +223,7 @@ public class DigipostApiMock implements HttpHandler {
 		this.requestsAndResponsesMap = requestsAndResponsesMap;
 		this.port = port;
 		this.server = Undertow.builder().addHttpListener(port, "localhost", new BlockingHandler(this)).build();
+		init();
 		server.start();
 		LOG.info("Evry SMS service mock running on port {}", port);
 		return this;
@@ -215,12 +234,19 @@ public class DigipostApiMock implements HttpHandler {
 		LOG.info("Shutting down Evry SMS service mock on port {}", port);
 	}
 
-	public void reset() {
-		receivedRequests.clear();
+	public void init() {
+		requestsAndResponsesMap.clear();
+		requestsAndResponsesMap.put(Method.GET_CONTENT, new RequestsAndResponses());
+		requestsAndResponsesMap.put(Method.GET_DOCUMENTS_EVENTS, new RequestsAndResponses());
+		requestsAndResponsesMap.put(Method.GET_DOCUMENT_STATUS, new RequestsAndResponses());
+		requestsAndResponsesMap.put(Method.GET_PRINT_KEY, new RequestsAndResponses());
+		requestsAndResponsesMap.put(Method.GET_SENDER_INFORMATION, new RequestsAndResponses());
+		requestsAndResponsesMap.put(Method.SEND_MULTIPART_MESSAGE, new RequestsAndResponses());
+		requestsAndResponsesMap.put(Method.MULTIPART_MESSAGE, new RequestsAndResponses(new MultipartRequestMatcher()));
 	}
 
-	public List<ApiServiceMock.ContentPart> getContentparts(String multipartString){
-		List<ApiServiceMock.ContentPart> contentParts = new ArrayList<>();
+	public List<ContentPart> getContentparts(String multipartString){
+		List<ContentPart> contentParts = new ArrayList<>();
 		String splitString = multipartString.substring(0, multipartString.indexOf("\r\n"));
 
 		String[] split = multipartString.split(splitString);
@@ -230,11 +256,182 @@ public class DigipostApiMock implements HttpHandler {
 				String contentType = bodyPart.substring(bodyPart.indexOf("Content-Type: ") + 14);
 				contentType = contentType.substring(0, contentType.indexOf("\r\n"));
 				if (!contentType.equals(MediaTypes.DIGIPOST_MEDIA_TYPE_V6)) {
-					contentParts.add(new ApiServiceMock.ContentPart(contentType));
+					contentParts.add(new ContentPart(contentType));
 				}
 			}
 		}
 
 		return contentParts;
+	}
+
+
+
+	public static class RequestMatcher {
+		public CloseableHttpResponse findResponse(String requestString) {
+			return null;
+		}
+	}
+
+	public static class RequestsAndResponses {
+		private final Queue<ResponseProducer> responseQueue = new ConcurrentLinkedQueue<>();
+		private final RequestMatcher requestMatcher;
+		private final Map<String, MockRequest> requestMap;
+
+		RequestsAndResponses() {
+			this(new RequestMatcher());
+		}
+
+		RequestsAndResponses(RequestMatcher requestMatcher) {
+			this.requestMatcher = requestMatcher;
+			this.requestMap = Collections.synchronizedMap(new LinkedHashMap<String, MockRequest>() {
+				@Override
+				protected boolean removeEldestEntry(Map.Entry<String, MockRequest> eldest) {
+					return size() > 100;
+				}
+			});
+		}
+
+		public void addExpectedResponse(CloseableHttpResponse response) {
+			responseQueue.add(new MockResponse(response));
+		}
+		public void addExpectedException(RuntimeException ex) {
+			responseQueue.add(new MockResponse(ex));
+		}
+
+		public CloseableHttpResponse getResponse() {
+			return getResponse("default");
+		}
+
+		public CloseableHttpResponse getResponse(String requestString) {
+			ResponseProducer response = responseQueue.poll();
+			if (response != null) {
+				return response.getResponse();
+			}
+
+			return requestMatcher.findResponse(requestString);
+		}
+
+		public void addRequest(MockRequest request) {
+			requestMap.put(request.getKey(), request);
+		}
+
+		public MockRequest getRequest(String requestKey) {
+			return requestMap.get(requestKey);
+		}
+
+		public Map<String, MockRequest> getRequests() {
+			return requestMap;
+		}
+	}
+
+	public static interface ResponseProducer {
+		CloseableHttpResponse getResponse();
+	}
+	public static class MockResponse implements ResponseProducer {
+
+		private CloseableHttpResponse response;
+		private RuntimeException exception;
+
+		public MockResponse(CloseableHttpResponse response) {
+			this.response = response;
+		}
+		public MockResponse(RuntimeException ex) {
+			this.exception = ex;
+		}
+
+		@Override
+		public CloseableHttpResponse getResponse() {
+			if (exception != null) {
+				throw exception;
+			}
+			return response;
+		}
+	}
+
+	public static class MockRequest {
+		private final String key;
+
+		public MockRequest(String key) {
+			this.key = key;
+		}
+
+		public String getKey() {
+			return key;
+		}
+	}
+
+	public static class MultipartRequestMatcher extends RequestMatcher {
+
+		public static CloseableHttpResponse DEFAULT_RESPONSE = getDefaultResponse();
+		public static RuntimeException CONNECTION_REFUSED = new RuntimeException(new ConnectException("Connection refused"));
+
+		public static final Map<String, CloseableHttpResponse> responses = new HashMap<>();
+		public static final Map<String, RuntimeException> errors = new HashMap<>();
+
+		public static CloseableHttpResponse getDefaultResponse(){
+			MessageDelivery messageDelivery = new MessageDelivery(UUID.randomUUID().toString(), Channel.DIGIPOST, COMPLETE, DateTime.now());
+
+			org.apache.commons.io.output.ByteArrayOutputStream bao = new org.apache.commons.io.output.ByteArrayOutputStream();
+			JAXB.marshal(messageDelivery, bao);
+
+			return MockfriendlyResponse.MockedResponseBuilder.create()
+					.status(SC_OK)
+					.entity(new ByteArrayEntity(bao.toByteArray()))
+					.build();
+		}
+
+
+		static {
+			responses.put("200:OK", DEFAULT_RESPONSE);
+			errors.put("CONNECTION_REFUSED", CONNECTION_REFUSED);
+		}
+
+		@Override
+		public CloseableHttpResponse findResponse(String requestString) {
+
+			if (responses.containsKey(requestString)) {
+				return responses.get(requestString);
+			} else if (errors.containsKey(requestString)) {
+				throw errors.get(requestString);
+			} else if (requestString.matches("^[0-9]{3}:(.)+")) {
+				String[] split = requestString.split(":");
+				if (ErrorCode.isKnown(split[1])) {
+					ErrorCode errorCode = ErrorCode.resolve(split[1]);
+					ErrorType translated = EnumUtils.getEnum(ErrorType.class, errorCode.getOverriddenErrorType().name());
+
+					org.apache.commons.io.output.ByteArrayOutputStream bao = new org.apache.commons.io.output.ByteArrayOutputStream();
+					JAXB.marshal(new ErrorMessage(translated != null ? translated : ErrorType.SERVER, errorCode.name(), "Generic error-message from digipost-api-client-mock"), bao);
+
+					return MockfriendlyResponse.MockedResponseBuilder.create().status(parseInt(split[0]))
+							.entity(new ByteArrayEntity(bao.toByteArray())).build();
+				} else {
+					throw new IllegalArgumentException("ErrorCode " + split[1] + " is unknown");
+				}
+			} else {
+				return DEFAULT_RESPONSE;
+			}
+
+		}
+	}
+
+	public static class DigipostRequest extends MockRequest {
+		public final Message message;
+		public final List<ContentPart> contentParts;
+
+		public DigipostRequest(Message message, List<ContentPart> contentParts) {
+			super(message.messageId);
+			this.message = message;
+			this.contentParts = contentParts;
+		}
+
+	}
+
+	public static class ContentPart {
+
+		public final String mediaType;
+
+		public ContentPart(String mediaType) {
+			this.mediaType = mediaType;
+		}
 	}
 }
