+++ src/main/java/no/digipost/api/client/util/FakeEncryptionKey.java
@@ -1,0 +1,50 @@
+/**
+ * Copyright (C) Posten Norge AS
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package no.digipost.api.client.util;
+
+import no.digipost.api.client.representations.EncryptionKey;
+import org.bouncycastle.openssl.jcajce.JcaPEMWriter;
+
+import java.io.ByteArrayOutputStream;
+import java.io.OutputStreamWriter;
+import java.io.Writer;
+import java.security.KeyPair;
+import java.security.KeyPairGenerator;
+
+public class FakeEncryptionKey {
+
+	public static EncryptionKey createFakeEncryptionKey() {
+		ByteArrayOutputStream baos = new ByteArrayOutputStream();
+		try (Writer osWriter = new OutputStreamWriter(baos); JcaPEMWriter writer = new JcaPEMWriter(osWriter)) {
+
+			KeyPairGenerator factory = KeyPairGenerator.getInstance("RSA");
+			factory.initialize(2048);
+			KeyPair keyPair = factory.generateKeyPair();
+
+			writer.writeObject(keyPair.getPublic());
+
+		} catch (Exception e) {
+			throw new RuntimeException("Failed creation of fake encryption key.", e);
+		}
+
+		EncryptionKey fakeKey = new EncryptionKey();
+		fakeKey.setKeyId("fake-hash");
+		fakeKey.setValue(new String(baos.toByteArray()));
+
+		return fakeKey;
+	}
+
+}
